[refactor] ReviewListContent의 데이터 패칭 로직 분리 및 컴포넌트 구조 개선
- ReviewListContent의 이름을 ReviewList로 변경했습니다.
- MainPage(서버 컴포넌트)에서 데이터를 요청한 뒤 BestReview 컴포넌트에서 카테고리를 선택해 ReviewList로 데이터를 주입하도록 변경했습니다.
