--- src/entities/reviews/model/useGetBestReviews.ts
@@ -1,10 +1,0 @@
-import {useSuspenseQuery} from '@tanstack/react-query';
-import {reviewQueryOptions} from './query-service';
-
-function useGetBestReviews() {
-  const {data} = useSuspenseQuery(reviewQueryOptions.best());
-
-  return data;
-}
-
-export default useGetBestReviews;
