+++ src/features/best-review/ui/BestReview.tsx
@@ -1,14 +1,18 @@
 'use client';
 
-import {CategoryBar} from '@/shared/ui/components';
-import {Suspense, useState} from 'react';
+import {useState} from 'react';
 import Link from 'next/link';
+import {ReviewList, BestReviewCategory, BestReviewMapped} from '@/entities/reviews';
+import {CategoryBar} from '@/shared/ui/components';
 import {LucideIcon} from '@/shared/ui/icons';
-import {BestReviewCategory} from '@/entities/reviews/model/types';
-import ReviewListContent from '../../../entities/reviews/ui/ReviewListContent';
 
-export default function BestReview() {
+type Props = {
+  reviews: BestReviewMapped;
+};
+
+export default function BestReview({reviews}: Props) {
   const [selectedCategory, setSelectedCategory] = useState<BestReviewCategory>('all');
+  const filteredReview = reviews[selectedCategory];
 
   const handleSelectCategory = (category: BestReviewCategory) => {
     setSelectedCategory(category);
@@ -18,10 +22,7 @@ export default function BestReview() {
     <section className=" flex flex-col items-center justify-center bg-boldBlue mt-16 py-12 md:px-8">
       <h4 className="text-white font-bold text-2xl mb-8 md:text-3xl">🔥 BEST 후기 🔥</h4>
       <CategoryBar selectedCategory={selectedCategory} onSelectCategory={handleSelectCategory} />
-      {/* Todo: Fallback UI 구현 */}
-      <Suspense fallback={<div>Loading...</div>}>
-        <ReviewListContent selectedCategory={selectedCategory} />
-      </Suspense>
+      <ReviewList filteredReview={filteredReview} />
       <Link
         href="/allreviews"
         className="flex items-center animate-bounce cursor-pointer text-white font-bold text-[20px]"
