+++ src/views/main/ui/MainPage.tsx
@@ -1,11 +1,14 @@
 import Hero from '@/widgets/hero';
 import {BestReview} from '@/features/best-review';
+import {getBestReviews} from '@/entities/reviews';
 
-export default function MainPage() {
+export default async function MainPage() {
+  const data = await getBestReviews();
+
   return (
     <section>
       <Hero />
-      <BestReview />
+      <BestReview reviews={data} />
     </section>
   );
 }
